Refactor import statement for EvidenceHistory component in reports page
